requirements.txt
Django==6.0.2
asgiref==3.11.1
pillow==12.1.1
sqlparse==0.5.5
tzdata==2025.3
requests==2.32.3
daphne==4.1.2
whitenoise==6.9.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
channels==4.2.0
groq==0.23.1
razorpay==1.4.2
twilio==9.6.3
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
numpy==1.26.4
pandas==2.2.3
matplotlib==3.8.4
scikit-learn==1.4.2
reportlab==4.0.7
